# swagger.json
{
    "openapi": "3.1.0",
    "info": {
        "description": "Authorization API for STEXS",
        "title": "Authorization API",
        "version": "1.0.0"
    },
    "host": "localhost",
    "basePath": "/auth/api/v1/",
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "components": {
        "schemas": {
            "AccessTokenResponse": {
                "type": "object",
                "properties": {
                    "access_token": {
                        "type": "string",
                        "description": "A token that grants access to the user's resources."
                    },
                    "refresh_token": {
                        "type": "string",
                        "description": "A token used to request a new access token without requiring the user's identifier and password."
                    },
                    "token_type": {
                        "type": "string",
                        "description": "The type of the access token (e.g., 'bearer')."
                    },
                    "expires": {
                        "type": "integer",
                        "format": "unix",
                        "description": "The expiration timestamp of the access token."
                    }
                },
                "example": {
                    "access_token": "jwt-token",
                    "refresh_token": "jwt-token",
                    "token_type": "bearer",
                    "expires": 1696457106
                }
            },
            "SuccessResponse": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "description": "A boolean indicator of the operation's success."
                    },
                    "message": {
                        "type": "string",
                        "description": "A human-readable message describing the result of the operation."
                    },
                    "timestamp": {
                        "type": "string",
                        "format":"ISO 8601",
                        "description": "The timestamp when the message was generated."
                    },
                    "data": {
                        "type": "object",
                        "description": "Additional relevant data returned with the response."
                    }
                },
                "example": {
                    "success": true,
                    "message": "Success message",
                    "timestamp": "2023-10-05T14:21:52.793Z",
                    "data": {}
                }
            }
        },
        "securitySchemes": {
            "Auth": {
                "type": "apiKey",
                "in": "header",
                "name": "Authorization",
                "description": "Bearer Token"
            }
        }
    },
    "paths": {
...
        "/sign-in": {
            "post": {
                "summary": "Sign In or Initialize Sign-In",
                "description": "Allows users to sign in or initialize the sign-in process.",
                "tags": [
                    "Sign In"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "identifier": {
                                        "type": "string",
                                        "description": "The email address or username associated with the account."
                                    },
                                    "password": {
                                        "type": "string",
                                        "description": "The password used for authentication and account access."
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successfully signed in or sign-in process initialized.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "type": "object",
                                            "properties": {
                                                "token": {
                                                    "type": "string",
                                                    "description": "A JSON Web Token (JWT) used to confirm the sign-in."
                                                },
                                                "types": {
                                                    "type": "array",
                                                    "description": "An array of supported Multi-Factor Authentication (MFA) types for the account."
                                                },
                                                "expires": {
                                                    "type": "integer",
                                                    "format": "unix",
                                                    "description": "The timestamp indicating the expiration time of the sign-in confirmation token."
                                                }
                                            }
                                        },
                                        {
                                            "$ref": "#/components/schemas/AccessTokenResponse"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/sign-out": {
            "post": {
                "summary": "Sign Out from the Current Session",
                "description": "Logs the user out from the current session.",
                "tags": ["Sign Out"],
                "security": [
                    {
                        "Auth": []
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successfully signed out from the current session"
                    }
                }
            }
        },
...
        "/sign-up": {
            "post": {
                "summary": "Create a New Account",
                "tags": ["Sign Up"],
                "description": "Register a new user account with the following required information.",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "username": {
                                        "type": "string",
                                        "description": "A unique username for the user, consisting of 1 to 20 Latin characters, and should not resemble an email address."
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "description": "The user's email address. Must be a valid email format (e.g., 'user@example.com')."
                                    },
                                    "password": {
                                        "type": "string",
                                        "description": "A secure password that must meet the following criteria: - At least 10 characters long - Contains at least one uppercase letter - Contains at least one lowercase letter - Contains at least one special character - Contains at least one number"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Account Successfully Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/verify": {
            "post": {
                "summary": "Verify Email Address",
                "tags": ["Verify Email"],
                "description": "Allows users to verify their email address after signing up. This endpoint is used to confirm the validity of the provided email address and activate the user's account.",
                "parameters": [
                    {
                        "name": "email",
                        "format": "email",
                        "in": "query",
                        "description": "The email address to be verified.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "token",
                        "in": "query",
                        "description": "The verification token sent to the user's email address.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Redirect to Sign In",
                        "headers": {
                            "Location": {
                                "description": "The URL to redirect to for sign-in with verification status.",
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/token": {
            "post": {
                "summary": "Refresh Access Token",
                "tags": ["Token"],
                "description": "Allows users to refresh their access token by exchanging a valid refresh token. Use this endpoint to obtain a new access token and refresh token pair, which can be used to continue accessing protected resources without the need for reauthentication.",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "refresh_token": {
                                        "type": "string",
                                        "format": "JWT",
                                        "description": "The user's valid refresh token to exchange for a new pair of access and refresh tokens."
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Token Refresh Successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AccessTokenResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/recovery": {
            "post": {
                "summary": "Initiate Password Recovery",
                "tags": ["Recovery"],
                "description": "Allows users to initiate the process of recovering their password. Use this endpoint to request a password reset link sent to the user's registered email address. This link will enable the user to create a new password.",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "description": "The user's registered email address for initiating password recovery."
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Password Recovery Requested",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref":"#/components/schemas/SuccessResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/user": {
            "get": {
                "summary": "Retrieve User Data",
                "tags": ["User"],
                "description": "Allows authenticated users to retrieve their own user data. This endpoint provides access to essential user information, including email, raw user metadata, creation timestamp, update timestamp, and user ID.",
                "security": [
                    {
                        "Auth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User Data Retrieved",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "id": {
                                            "type": "string",
                                            "format": "uuid",
                                            "description": "A unique UUID identifier for the user's account."
                                        },
                                        "email": {
                                            "type": "string",
                                            "format": "email",
                                            "description": "The user's registered email address."
                                        },
                                        "raw_user_meta_data": {
                                            "type": "object",
                                            "description": "Raw user metadata associated with the user's account."
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "ISO 8601",
                                            "description": "Timestamp indicating when the user's account was created."
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "ISO 8601",
                                            "description": "Timestamp indicating when the user's account was last updated."
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/mfa": {
            "get": {
                "summary": "Check MFA Status",
                "tags": ["Multi-Factor Authentication (MFA)"],
                "description": "Allows authenticated users to check the status of their Multi-Factor Authentication (MFA) flows. This endpoint provides information on whether MFA methods like email and TOTP (Time-Based One-Time Password) are enabled or disabled for the user's account.",
                "security": [
                    {
                        "Auth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "MFA Status Retrieval Successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "email": {
                                            "type": "boolean",
                                            "description": "Indicates whether email-based MFA is enabled for the user's account."
                                        },
                                        "totp": {
                                            "type": "boolean",
                                            "description": "Indicates whether TOTP (Time-Based One-Time Password) MFA is enabled for the user's account."
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/mfa/verify": {
            "post": {
                "summary": "Verify MFA",
                "tags": ["Multi-Factor Authentication (MFA)"],
                "description": "Allows users to verify the setup of MFA for their account. Users need to provide the type of MFA used and the verification code to confirm the setup of MFA for their account.",
                "security": [
                    {
                        "Auth": []
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "type": {
                                        "type": "string",
                                        "description": "The type of MFA being verified (e.g., 'totp')."
                                    },
                                    "code": {
                                        "type": "string",
                                        "description": "The verification code for the specified MFA type."
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "MFA Setup Verification Successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}